Only delete temporary bundles

+++ src/main/java/uk/org/taverna/databundle/DataBundle.java
@@ -1,14 +1,16 @@
 package uk.org.taverna.databundle;
 
+import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
 public class DataBundle implements AutoCloseable {
 
 	private final Path root;
-
-	public DataBundle(Path root) {
+	private boolean deleteOnClose;
+	public DataBundle(Path root, boolean deleteOnClose) {
 		this.root = root;
+		this.deleteOnClose = deleteOnClose;
 	}
 
 	public Path getRoot() {
@@ -16,9 +18,13 @@ public class DataBundle implements AutoCloseable {
 	}
 
 	@Override
-	public void close() throws Exception {
-		Path orig = DataBundles.closeDataBundle(this);
-		Files.deleteIfExists(orig);
+	public void close() throws IOException  {
+		if (getRoot().getFileSystem().isOpen()) {
+			Path orig = DataBundles.closeDataBundle(this);
+			if (deleteOnClose) {
+				Files.deleteIfExists(orig);
+			}
+		}
 	}
 
 }
